+++ resources/js/Components/Checkbox.vue
@@ -29,6 +29,6 @@ const proxyChecked = computed({
         type="checkbox"
         :value="value"
         v-model="proxyChecked"
-        class="rounded border-gray-300 text-indigo-600 shadow-sm focus:ring-indigo-500"
+        class="rounded border-brand-mist text-brand-blue shadow-sm focus:ring-brand-orange/40"
     />
 </template>