+++ resources/views/app.blade.php
@@ -4,13 +4,16 @@
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1">
 
-        <title inertia>{{ config('app.name', 'Laravel') }}</title>
+        <title inertia>{{ config('app.name', 'TeamForge') }}</title>
 
-        <!-- Fonts -->
+        <link rel="icon" href="/icon_small_blue.svg" type="image/svg+xml">
+
         <link rel="preconnect" href="https://fonts.bunny.net">
-        <link href="https://fonts.bunny.net/css?family=figtree:400,500,600&display=swap" rel="stylesheet" />
+        <link
+            href="https://fonts.bunny.net/css?family=plus-jakarta-sans:400,500,600,700&display=swap"
+            rel="stylesheet"
+        />
 
-        <!-- Scripts -->
         @routes
         @vite(['resources/js/app.js', "resources/js/Pages/{$page['component']}.vue"])
         @inertiaHead